docker and fix smol stuff
