+++ docker-compose.yml
@@ -1,0 +1,17 @@
+services:
+  frontend:
+    build: ./ui
+    container_name: sparkflow-frontend
+    ports:
+      - "3000:3000"
+    develop:
+      watch:
+        - action: sync
+          path: ./ui
+          target: /app
+          ignore:
+            - node_modules/
+            - .next/
+            - .git/
+        - action: rebuild
+          path: ./ui/package.json
